requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Utilities
tqdm>=4.65.0

# Testing
pytest>=7.3.0
